requirements.txt
fastapi
uvicorn[standard]
python-dotenv
supabase
openai>=1.0.0
